requirements.txt
pandas==1.3.5
requests==2.26.0
seaborn==0.11.2
streamlit==1.2.0
google-auth==2.3.3
gspread=5.0.0